requirements.txt
MySQL-python==1.2.5
web.py==0.38
pyOpenSSL==16.2.0
python-telegram-bot==9.0.0
